Create next_prev.html
I though it might be nice to have a next_prev.html include that could be included at the top and bottom of pages. The next_prev.html include uses the nav_order page property to find the next or previous page in the sequence and then makes these available as links. This could be added as desired to the default layout like so:
```
{% include next_prev.html %}
{{content}}
{% include next_prev.html %}