Write search-data.json from heredoc
When file `assets/js/zzzz-search-data.json` is created with rake,
invalid liquid is created. The problem is unescaped single quotes
in the string from which the file is written. To fix this bug, we
could simply escape these single quotes. However, this multiline
string already contains many escaped single quotes. To avoid such
issues in the future, the file is written from a heredoc instead.

Resolves #488