Fix header margins for AsciiDoc
The jekyll-asciidoc[1] plugin generates div.sect(𝑛 − 1) elements
around headings of type h𝑛, that enclose all the heading and all the
content after it until the next heading of greater or equal rank.

This means that headings are _always_ first children in AsciiDoc
output, which meant the wrong margins were applied to most headings.
To fix this, only reduce the margin of first-child headings nested
directly below the .main-content element, and headings nested directly
below AsciiDoc .sect𝑛 elements that are themselves first children.

[1]: https://github.com/asciidoctor/jekyll-asciidoc