Fix Table of Contents styling for AsciiDoc
AsciiDoc (with the jekyll-asciidoc plugin), unfortunately does not use
an HTML heading element for the heading of its table of contents.

In terms of document structure, AsciiDoc's table of contents heading
is always equivalent to an h2, so style it the same.

Unfortunately, the only way to target the heading is using its ID, so
we have to disable the corresponding stylelint warnings.