Update nav.html
The possibility of using the `title` as a default value for `nav_order` requires page grouping
(because of the lack of support in Jekyll-Liquid for sorting by the value of an expression).
It appears that extracting a sorted list of pages from a sorted list of groups is quadratic
in the number of groups: it can take 35 seconds for 130 single-value groups.

This update to nav.html circumvents this source of inefficiency by generating the navigation
directly from the sorted page groups.

Building the same site with 130 pages now takes only about 6 seconds with Jekyll v4.2.2.

In the previous commit, distinguishing between numbers and strings didn't work.
This commit has been partially tested, locally.