Fix excluded vendor
Fix #973

Update the array of excluded directories and files to:
- those that Jekyll excludes by default, and
- the others that the theme previously excluded, but omitting `vendor` and ones that are no longer in the repo.

Excluding `vendor` was preventing HTML compression of the theme docs, since the layout used for compression is in `_layouts/vendor/compress.html`.

The change from one-line to multi-line YAML layout for the array of excluded files enhances readability. The addition of explanatory comments supports maintenance.