Add configuration key to load a local version of mermaid
In some cases the usage of jsDelivr might not be possible for technical or
compliance reasons.

This commit adds a second way to include the mermaid lib by specifying a path in
the mermaid config. This way a local version of the lib can be used.

It should be fully backwards compatible, not requiring any action by users that
already include the lib from the CDN.