Fix empty collections and excluded pages
Empty collections are now treated specially, because they do not contain any nav-list-items, and a nav_split can contain any number of them.

Pages are implicitly excluded from the main navigation when their parents or grandparents are explicitly excluded. Testing whether the nav-list-link for the current page is in the site-nav detects all excluded pages.