Show navigation to excluded pages
See discussion #1566 for motivation.

Currently (v0.10.0) the cached site-nav used in the sidebar contains links only to pages that are not excluded from the main navigation.
When an excluded page is selected, the navigation path to any ancestors disappears.

This draft PR makes the cached site-nav used in the sidebar contain links to all pages, but adds the `nav-exclude` class to the HTML of links to excluded pages.
The CSS for the navigation then suppresses the display of those links, and the JS scrolls the navigation to the closest non-excluded ancestor link.

Preliminary testing with the Just-the-Docs-Tests repo was successful with JS enabled.
The generated page-specific CSS will need adjustment to give the same navigation when JS is disabled.
Rigorous testing will be needed to ensure that this should be a non-breaking change for non-customised sites.