Change aria-pressed to aria-expanded for expanable content
This PR changes aria-pressed to aria-expanded on buttons toggeling expanable content.

Aria-pressed is used for e.g. switch buttons, a screen reader will announce "on" or "off" with the button.

For expandable content like menus the correct attribute is aria-expanded. The screen reader will announce the accessible name if the button adding the state expanded or collapsed when the button gets forcus.