simplify aria-label for buttons with aria-expanded
When aria-expanded is used there is no need to add to the aria-label the text “toggle", because the aria-expanded already gives that info and the state of the toggle.
So: “Menu”, instead of “Toggle menu”.
The screen reader will announce the info as:
Button Menu collapsed
and
Button Menu expanded

Also it is for voice recognition software advisable to start the aria-label with the text a user may guess the button stands for. (This is best practice, not WCAG)